extractors/computermetrics/pyproject.toml
[project]
name = "computermetrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil>=7.2.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
